% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_name.R
\name{fix_name}
\alias{fix_name}
\title{fix_name}
\usage{
fix_name(name)
}
\arguments{
\item{name}{}
}
\value{
name Character. Correctly spelled electorate name.
}
\description{
Fix an incorrectly title, spelled or misnamed electorate name.
}
\examples{
fix_name('Occonor')
}
