% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_name.R
\name{validate_name}
\alias{validate_name}
\title{validate_name}
\usage{
validate_name(name)
}
\arguments{
\item{name}{The name of an electorate for the 2018 election.}
}
\value{
Binary indicating whether the name is valid.
}
\description{
Check whether the supplied name is a valid name for a Commonwealth Electoral
Division at the 2018 Federal Election.
}
\examples{
validate_name('Bean')
}
